+++ bpx/__init__.py
@@ -1,7 +1,7 @@
 """BPX schema and parsers"""
 # flake8: noqa F401
 
-__version__ = "0.3.1"
+__version__ = "0.4.0"
 
 from .interpolated_table import InterpolatedTable
 from .expression_parser import ExpressionParser